model/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.11.6
six==1.17.0
tqdm==4.67.1
Werkzeug==3.1.3
